requirements.txt
# DEIM Requirements
# Python 3.8+ required

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.7.0
pillow>=9.0.0
pyyaml>=6.0

# Visualization
supervision>=0.17.0

# Data handling
pandas>=1.5.0
scipy>=1.9.0
matplotlib>=3.5.0

# Development tools (optional)
ruff>=0.1.0
pyright>=1.1.0

# Additional dependencies for training
tqdm>=4.65.0
tensorboard>=2.11.0